+++ src/Template/PHPFile/staticProperties.php
@@ -6,16 +6,15 @@
  *
  * For the full copyright and license information, read the LICENSE file that was distributed with this source code.
  */
-/**
- * @var $table \Database2Code\Struct\Table
- */
 
+/** @var $table \Database2Code\Struct\Table */
+/** @var $config \Database2Code\Output\OutputConfig */
 
-if(!function_exists('staticProperties__getColumnMethods')) {
-    function staticProperties__getColumnMethods(\Database2Code\Struct\Column $column)
+if(!function_exists('PHPFile__staticProperties__getColumnConstant')) {
+    function PHPFile__staticProperties__getColumnConstant(\Database2Code\Struct\Column $column)
     {
         return <<<END
-    
+
     public static \${$column->getName()} = '{$column->getName()}';
 END;
     }
@@ -23,13 +22,21 @@ END;
 
 $properties = '';
 foreach ($table->getColumns() as $column) {
-    $properties .= staticProperties__getColumnMethods($column).PHP_EOL;
+    $properties .= PHPFile__staticProperties__getColumnConstant($column).PHP_EOL;
+}
+
+if($config->hasNamespace()) {
+    $namespace = "\nnamespace {$config->getNamespace()};\n";
+} else {
+    $namespace = '';
 }
 
 return <<<END
 <?php
-
+$namespace
 class {$table->getName()} {
+
+    public static \$from = '{$table->getName()}';
 $properties
 }
 END;