+++ ChangeLog.md
@@ -5,3 +5,13 @@ All notable changes of the Database2Code release series are documented in this f
 ## [0.9.0] - 2018-06-30
 
 * Initial Release
+
+## [0.10.0] - 2019-03-16
+
+* Added third output template file "constants"
+* Added more mysql column types
+* Output-template can now access config values
+* Added console help text
+* Improved error handling
+* Added more unittests
+* Cleaned up code smells